+++ docs/CHANGES.md
@@ -3,7 +3,9 @@
 ## 0.10.1 (Unreleased)
 
 - #524 Added `auto_join_on_invite` parameter for automatically joining chatrooms. [ben]
-- FIX: A chatroom invite might come from someone not in your roster list. [ben]
+- #521 Not sending presence when connecting after disconnection. [jcbrand]
+- #536 Presence not sent out (in cases where it should) after page refresh. [jcbrand]
+- A chatroom invite might come from someone not in your roster list. [ben]
 
 ## 0.10.0 (2015-11-05)